verifier/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
azure-identity==1.17.1
azure-storage-blob==12.22.0
